requirements.txt
python-dotenv
vertica-python
